Commit loginpage with animation
