+++ src/Component/Blob.css
@@ -6,7 +6,7 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    background: #e4e2df;
+    background: #ffffff;
 }
 
 .blob {
@@ -20,20 +20,21 @@
 }
 
 .blob-1 {
-    top: 100px;
-    left: -5%;
+    top: 10%;
+    right: 10%;
     animation-delay: 0s;
 }
 
 .blob-2 {
-    bottom: 200px;
-    right: -100;
+    top: 20%;
+
     animation-delay: 10s;
 }
 
 .blob-3 {
-    top: 0%;
 
+    top: 5%;
+    left: 15%;
     transform: translateX(-50%);
     animation-delay: 20s;
 }
@@ -55,17 +56,4 @@
         border-radius: 60% 30% 40% 70%;
         transform: scale(1.3) translate(10px, -10px);
     }
-}
-
-.blob-content {
-    position: relative;
-    z-index: 10;
-    max-width: 700px;
-    padding: 2rem;
-    background-color: rgba(255, 255, 255, 0.8);
-    backdrop-filter: blur(12px);
-    border-radius: 24px;
-    text-align: center;
-    color: #0064ce;
-    box-shadow: 0 0 40px rgb(255, 255, 255);
 }
