+++ src/Component/Launcher.css
@@ -1,0 +1,50 @@
+.image-layer{position: relative}
+.form-image-main{width: 100%}
+.form-image{ position: absolute;
+left: 0;
+width: 100%}
+.form-image-Eye{ position: absolute;
+    left: 0;
+    height: 40px;
+    border-radius: 50%;
+    z-index: -1;
+    width: 40px}
+
+.coin{
+    animation: scale-down 3s ease-in-out alternate infinite;
+}
+.spring{
+    animation: scale-down 3s ease-in-out alternate infinite;
+}
+.dots{
+    animation: scale-up 3s ease-in-out alternate infinite;
+}
+.rocket{
+    animation: up-down 3s ease-in-out alternate infinite;
+}
+.cloud{
+    animation: left-right 3s ease-in-out alternate infinite;
+}
+.stars{
+    animation: scale-down 3s ease-in-out alternate infinite;
+}
+@keyframes scale-down {
+    to{
+        transform: scale(0.95);
+    }
+}
+@keyframes scale-up {
+    to{
+        transform: scale(1.05);
+    }
+}
+@keyframes up-down {
+    to{
+        transform: translateY(10px);
+    }
+}
+@keyframes left-right {
+    to{
+        transform: translateX(10px);
+    }
+}
