+++ src/Component/Launcher.tsx
@@ -1,0 +1,16 @@
+import './Launcher.css';
+
+const Launcher = () => (
+    <div className="image-layer">
+        <img src="src/assets/white-outline.png" className="form-image-main" alt="Image 1" />
+        <img src="src/assets/dots.png" className="form-image dots" alt="Image 2" />
+        <img src="src/assets/coin.png" className="form-image coin" alt="Image 3" />
+        <img src="src/assets/spring.png" className="form-image spring" alt="Image 4" />
+        <img src="src/assets/rocket.png" className="form-image rocket" alt="Image 5" />
+        <img src="src/assets/cloud.png" className="form-image cloud" alt="Image 6" />
+        <img src="src/assets/stars.png" className="form-image stars" alt="Image 7" />
+
+    </div>
+);
+
+export default Launcher;
