--- src/HomePage.css
@@ -1,36 +1,0 @@
-.home-container {
-    text-align: center;
-    padding: 2rem;
-    background-color: rgba(255, 255, 255, 0.2);
-    backdrop-filter: blur(10px);
-    border-radius: 24px;
-    box-shadow: 0 0 40px rgba(0, 0, 0, 0.1);
-}
-.home-container h1 {
-    font-size: 3rem;
-    margin-bottom: 1rem;
-    color: #ffffff;
-}
-
-.home-container p {
-    font-size: 1.2rem;
-
-    color: #f0f0f0;
-    margin-bottom: 1rem;
-}
-
-.github-button {
-    margin-top: 1.5rem;
-    padding: 12px 24px;
-    font-size: 1rem;
-    color: white;
-    background-color: #24292f;
-    border: none;
-    border-radius: 12px;
-    cursor: pointer;
-    transition: background 0.3s ease;
-}
-
-.github-button:hover {
-    background-color: #3c4147;
-}
