+++ src/LoginPage.css
@@ -1,0 +1,97 @@
+.home-container {
+    text-align: center;
+    padding: 2rem;
+    background-color: rgba(255, 255, 255, 0.2);
+    backdrop-filter: blur(10px);
+    border-radius: 24px;
+    box-shadow: 0 0 40px rgba(0, 0, 0, 0.1);
+}
+.grid-item h1 {
+    font-family: "Roboto Slab", system-ui;
+    font-size: 3rem;
+    margin-bottom: 1rem;
+    color: rgba(27, 30, 57, 0.8);
+}
+.grid-item h1 span{
+    font-weight: bolder;
+}
+
+.grid-item p {
+    font-size: 1.2rem;
+    color: rgba(27, 30, 57, 0.8);
+    margin-bottom: 1rem;
+}
+
+.github-button {
+    margin-top: 1.5rem;
+    padding: 12px 24px;
+    font-size: 1rem;
+    color: white;
+    background-color: #ed2f6c;
+    border: none #ed2f6c;
+    border-radius: 12px;
+    cursor: pointer;
+    transition: background 0.3s ease;
+}
+
+.github-button:hover {
+    background-color: #ffffff;
+    border: 0.2px solid #ed2f6c;
+    border-radius:12px;
+
+    color: #ed2f6c;
+}
+
+
+
+.grid-item {
+    width: 100%;
+    color: white;
+    padding: 20px;
+    border-radius: 8px;
+    box-sizing: border-box;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+}
+
+
+.blob-content {
+    position: relative;
+    z-index: 10;
+    max-width: 75%;
+    padding: 2rem;
+    background-color: rgba(255, 255, 255, 0.8);
+    backdrop-filter: blur(12px);
+    border-radius: 24px;
+    text-align: center;
+    color: #0064ce;
+    box-shadow: 0 0 40px rgb(255, 255, 255);
+}
+.grid-container {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    height: auto;
+    width: 100%;
+    gap: 20px;
+    padding: 20px;
+    align-items: center;
+}
+
+
+@media (max-width: 800px) {
+    .grid-container {
+        grid-template-columns: 1fr;
+    }
+
+    .grid-item:first-child {
+        order: 2;
+    }
+
+    .grid-item:last-child {
+        order: 1;
+    }
+}
+.logo{
+    border-radius: 50%;
+}
