+++ src/index.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@100..900&display=swap');
+
 :root {
   font-family: system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
@@ -77,3 +79,25 @@ button:focus-visible {
     background-color: #f9f9f9;
   }
 }
+
+@font-face {
+  font-family: 'Mint Round';
+  src: url('../public/fonts/MintRound.woff2') format('woff2');
+  font-weight: normal;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: 'Mint Round-Inline';
+  src: url('../public/fonts/MintRoundInline.woff2') format('woff2');
+  font-weight: normal;
+  font-style: normal;
+}
+
+@font-face {
+  font-family: 'Mint Round - Outline';
+  src: url('../public/fonts/MintRoundOutline.woff2') format('woff2');
+  font-weight: normal;
+  font-style: normal;
+}
+
